implementado deletar correntista

+++ src/app/components/correntista/correntista.component.ts
@@ -1,7 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, SimpleChange } from '@angular/core';
 import { response } from 'express';
 import { CorrentistaService } from 'src/app/services/correntista.service';
 import { isTemplateExpression } from 'typescript';
+import { ModalDismissReasons, NgbModal, NgbModalOptions } from '@ng-bootstrap/ng-bootstrap';
+import { ModalComponent } from '../modal/modal.component';
 
 @Component({
   selector: 'app-correntista',
@@ -18,7 +20,7 @@ export class CorrentistaComponent implements OnInit {
   
   constructor(
     private correntistaService: CorrentistaService,
-    
+    private modalService: NgbModal, 
     
   ) { }
 
@@ -50,6 +52,8 @@ export class CorrentistaComponent implements OnInit {
       response => {
         console.log(response);
         this.exibirCorrentistas();
+        this.cpf=''
+        this.nome=''
       },
       error => {
         console.log(error);
@@ -59,13 +63,15 @@ export class CorrentistaComponent implements OnInit {
 delete(idConta:any) {
   
   console.log("id", idConta);
-
-  this.correntistaService.delete(idConta)
-  .subscribe((data:any)=>{
-    console.log(data);
-    console.log("correntista deletado");  
+  if(confirm('Tem certeza que deseja deletar o correntista?')){
+    this.correntistaService.delete(idConta)
+    .subscribe((data:any)=>{
+      console.log(data);
+      console.log("correntista deletado");
+      this.exibirCorrentistas();  
   });
-  
+  }
 }
+
 
 }
